+++ CHANGELOG.md
@@ -1,11 +1,21 @@
-## 0.0.26-dev4
+
+## 0.1.1-dev
+
+### Enhancements
+
+* **Migrate Slack Source Connector to V2**
+
+## 0.1.0
 
 ### Enhancements
 
 * **Move default API URL parameter value to serverless API**
 * **Add check that access config always wrapped in Secret**
 * **Add togetherai embedder support**
-* **Migrate Slack Source Connector to V2**
+* **Refactor sqlite and postgres to be distinct connectors to support better input validation**
+* **Added MongoDB source V2 connector**
+* **Support optional access configs on connection configs**
+* **Refactor databricks into distinct connectors based on auth type**
 
 ### Fixes